archive/requirements-freeze.txt
# Requirements congelados para reproducibilidade - Gerado em 2025-09-30
# Execute: pip install -r requirements-freeze.txt para ambiente idêntico

# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning
lightgbm>=3.3.0
xgboost>=1.7.0
optuna>=3.0.0

# Notebooks & Visualization
jupyter>=1.0.0
ipykernel>=6.0.0
plotly>=5.0.0

# Model Explainability  
shap>=0.41.0

# Utilities
pyyaml>=6.0
tqdm>=4.60.0
joblib>=1.2.0

# Testing (development)
pytest>=7.0.0

# Note: Para gerar versões exatas, execute:
# pip freeze > requirements-freeze-exact.txt
# após configurar ambiente completo